requirements-openvoice.txt
librosa
faster-whisper
pydub
wavmark
numpy
eng_to_ipa
inflect
unidecode
whisper-timestamped
openai
python-dotenv
pypinyin
cn2an
jieba
gradio
langid
datasets
huggingface-hub
scipy
difussers
transformers
accelerate
sentencepiece
